+++ PizzaStoreApp/Pages/Management/Products/Index.cshtml
@@ -12,7 +12,7 @@
     var session = @HttpContextAccessor.HttpContext.Session;
     Account login = SessionExtension.GetLoginUser(session);
 }
-<h1>Index</h1>
+<h1>Pizza</h1>
 
 <p>
     <a asp-page="Create">Create New</a>
